Make the objects behave as transforms directly
Remove the .stream() static methods and instead let TextEncoder and
TextDecoder objects be used directly in .pipeThrough(). This is done by
delegating to a TransformStream object for the "readable" and "writable"
attributes. encode() and decode() methods now throw if the "readable"
or "writable" are locked.

This is an almost complete rewrite. The core algorithms have not changed
but input handling is more robust.